\name{lmObj}
\alias{lmObj}
\docType{data}
\title{
Example of a LowMACA object
}
\description{
An object of class LowMACA of the alignment and mapping of the homeobox domain. 
It is the example used in the vignette.
}
\usage{data("lmObj")}
\format{
An object of class LowMACA
}
\source{
Created by LowMACA package
}
\examples{
#Load lmObj and show its structure
data(lmObj)
str(lmObj)
}
\keyword{datasets}
